requirements.txt
gensim
networkx
nodevectors
numpy
pandas
pretty_midi
PyYAML
scikit-learn
tensorflow
tqdm
